Revert "(PC-26392) refactor(Search): split search in Landing & Result"
