+++ src/features/search/components/SearchHeader/SearchHeader.native.test.tsx
@@ -3,6 +3,7 @@ import { v4 as uuidv4 } from 'uuid'
 
 import { SearchHeader } from 'features/search/components/SearchHeader/SearchHeader'
 import * as useFilterCountAPI from 'features/search/helpers/useFilterCount/useFilterCount'
+import { SearchView } from 'features/search/types'
 import * as useFeatureFlagAPI from 'libs/firebase/firestore/featureFlags/useFeatureFlag'
 import { render, screen, waitFor, within } from 'tests/utils'
 
@@ -44,7 +45,7 @@ const useFeatureFlagSpy = jest.spyOn(useFeatureFlagAPI, 'useFeatureFlag').mockRe
 
 describe('SearchHeader component', () => {
   it('should render SearchHeader', async () => {
-    renderSearchHeader({ shouldDisplaySubtitle: true, isDesktopViewport: false })
+    renderSearchHeader(false, SearchView.Landing)
 
     await screen.findByText('Rechercher')
 
@@ -53,7 +54,7 @@ describe('SearchHeader component', () => {
 
   it('should show LocationWidget when ENABLE_APP_LOCATION featureFlag is on and when isDesktopViewport is false and SearchView is Landing', async () => {
     useFeatureFlagSpy.mockReturnValueOnce(true)
-    renderSearchHeader({ shouldDisplaySubtitle: true, isDesktopViewport: false })
+    renderSearchHeader(false, SearchView.Landing)
 
     await waitFor(() => {
       const insideLocationWidget = within(screen.getByTestId('InsideLocationWidget'))
@@ -64,7 +65,7 @@ describe('SearchHeader component', () => {
 
   it('should not show LocationWidget when ENABLE_APP_LOCATION featureFlag is off and when isDesktopViewport is false and SearchView is Landing', async () => {
     useFeatureFlagSpy.mockReturnValueOnce(false)
-    renderSearchHeader({ shouldDisplaySubtitle: true, isDesktopViewport: false })
+    renderSearchHeader(false, SearchView.Landing)
 
     await waitFor(() => {
       const insideLocationWidget = within(screen.getByTestId('InsideLocationWidget'))
@@ -75,7 +76,7 @@ describe('SearchHeader component', () => {
 
   it('should not show LocationWidget when ENABLE_APP_LOCATION featureFlag is on and when isDesktopViewport is false and searchView is not landing', async () => {
     useFeatureFlagSpy.mockReturnValueOnce(true)
-    renderSearchHeader({ shouldDisplaySubtitle: false, isDesktopViewport: false })
+    renderSearchHeader(false, SearchView.Results)
 
     await waitFor(() => {
       const insideLocationWidget = within(screen.getByTestId('InsideLocationWidget'))
@@ -84,9 +85,9 @@ describe('SearchHeader component', () => {
     })
   })
 
-  it('should show SearchLocationWidget when ENABLE_APP_LOCATION featureFlag is on and when isDesktopViewport is false', async () => {
+  it('should show SearchLocationWidget when ENABLE_APP_LOCATION featureFlag is on and when isDesktopViewport is false and SearchView is Landing', async () => {
     useFeatureFlagSpy.mockReturnValueOnce(true)
-    renderSearchHeader({ shouldDisplaySubtitle: true, isDesktopViewport: false })
+    renderSearchHeader(false, SearchView.Landing)
 
     await waitFor(() => {
       const searchHeaderTitleContainer = within(screen.getByTestId('SearchHeaderTitleContainer'))
@@ -95,9 +96,9 @@ describe('SearchHeader component', () => {
     })
   })
 
-  it('should not show LocationWidget when ENABLE_APP_LOCATION featureFlag is on and when isDesktopViewport is true', async () => {
+  it('should not show LocationWidget when ENABLE_APP_LOCATION featureFlag is on and when isDesktopViewport is true and SearchView is Landing', async () => {
     useFeatureFlagSpy.mockReturnValueOnce(true)
-    renderSearchHeader({ shouldDisplaySubtitle: true, isDesktopViewport: true })
+    renderSearchHeader(true, SearchView.Landing)
 
     await waitFor(() => {
       const insideLocationWidget = within(screen.getByTestId('InsideLocationWidget'))
@@ -106,9 +107,20 @@ describe('SearchHeader component', () => {
     })
   })
 
-  it('should show SearchLocationWidget when ENABLE_APP_LOCATION featureFlag is on and when isDesktopViewport is true', async () => {
+  it('should show SearchLocationWidget when ENABLE_APP_LOCATION featureFlag is on and when isDesktopViewport is true and SearchView is Landing', async () => {
     useFeatureFlagSpy.mockReturnValueOnce(true)
-    renderSearchHeader({ shouldDisplaySubtitle: true, isDesktopViewport: true })
+    renderSearchHeader(true, SearchView.Landing)
+
+    await waitFor(() => {
+      const searchHeaderTitleContainer = within(screen.getByTestId('SearchHeaderTitleContainer'))
+
+      expect(searchHeaderTitleContainer.queryByText('Me localiser')).toBeOnTheScreen()
+    })
+  })
+
+  it('should show SearchLocationWidget when ENABLE_APP_LOCATION featureFlag is on and when isDesktopViewport is true and SearchView is not landing', async () => {
+    useFeatureFlagSpy.mockReturnValueOnce(true)
+    renderSearchHeader(true, SearchView.Results)
 
     await waitFor(() => {
       const searchHeaderTitleContainer = within(screen.getByTestId('SearchHeaderTitleContainer'))
@@ -119,7 +131,7 @@ describe('SearchHeader component', () => {
 
   it('should not show SearchLocationWidget when ENABLE_APP_LOCATION featureFlag is off and when isDesktopViewport is true', async () => {
     useFeatureFlagSpy.mockReturnValueOnce(false)
-    renderSearchHeader({ shouldDisplaySubtitle: true, isDesktopViewport: true })
+    renderSearchHeader(true)
 
     await waitFor(() => {
       const searchHeaderTitleContainer = within(screen.getByTestId('SearchHeaderTitleContainer'))
@@ -129,18 +141,13 @@ describe('SearchHeader component', () => {
   })
 })
 
-interface RenderSearchHeaderProps {
-  shouldDisplaySubtitle: boolean
-  isDesktopViewport?: boolean
-}
-
-function renderSearchHeader({ shouldDisplaySubtitle, isDesktopViewport }: RenderSearchHeaderProps) {
+function renderSearchHeader(isDesktopViewport?: boolean, searchView?: SearchView) {
   const searchInputID = uuidv4()
 
   return render(
     <SearchHeader
       searchInputID={searchInputID}
-      shouldDisplaySubtitle={shouldDisplaySubtitle}
+      searchView={searchView}
       addSearchHistory={jest.fn()}
       searchInHistory={jest.fn()}
     />,
