+++ src/features/search/components/SearchResults/SearchResults.native.test.tsx
@@ -3,7 +3,7 @@ import { v4 as uuidv4 } from 'uuid'
 
 import { SearchGroupNameEnumv2 } from 'api/gen'
 import { useAuthContext } from 'features/auth/context/AuthContext'
-import { SearchResultsContent } from 'features/search/components/SearchResultsContent/SearchResultsContent'
+import { SearchResults } from 'features/search/components/SearchResults/SearchResults'
 import { initialSearchState } from 'features/search/context/reducer'
 import { MAX_RADIUS } from 'features/search/helpers/reducer.helpers'
 import { LocationFilter, SearchState, SearchView, UserData } from 'features/search/types'
@@ -122,7 +122,7 @@ const Kourou: SuggestedPlace = {
 
 jest.useFakeTimers({ legacyFakeTimers: true })
 
-describe('SearchResultsContent component', () => {
+describe('SearchResults component', () => {
   beforeAll(() => {
     mockHits = []
     mockNbHits = 0
@@ -139,7 +139,7 @@ describe('SearchResultsContent component', () => {
 
   it('should render correctly', async () => {
     jest.advanceTimersByTime(2000)
-    render(<SearchResultsContent />)
+    render(<SearchResults />)
     await act(async () => {})
 
     expect(screen).toMatchSnapshot()
@@ -157,7 +157,7 @@ describe('SearchResultsContent component', () => {
       },
     })
 
-    render(<SearchResultsContent />)
+    render(<SearchResults />)
 
     const flashList = screen.getByTestId('searchResultsFlashlist')
 
@@ -183,14 +183,14 @@ describe('SearchResultsContent component', () => {
 
   describe('Category filter', () => {
     it('should display category filter button', async () => {
-      render(<SearchResultsContent />)
+      render(<SearchResults />)
       await act(async () => {})
 
       expect(screen.getByTestId('Catégories')).toBeOnTheScreen()
     })
 
     it('should open the categories filter modal when pressing the category button', async () => {
-      render(<SearchResultsContent />)
+      render(<SearchResults />)
       const categoryButton = screen.getByTestId('Catégories')
 
       await act(async () => {
@@ -207,7 +207,7 @@ describe('SearchResultsContent component', () => {
         ...mockSearchState,
         offerCategories: [SearchGroupNameEnumv2.CD_VINYLE_MUSIQUE_EN_LIGNE],
       }
-      render(<SearchResultsContent />)
+      render(<SearchResults />)
       await act(async () => {})
 
       const categoryButtonIcon = screen.getByTestId('categoryButtonIcon')
@@ -225,14 +225,14 @@ describe('SearchResultsContent component', () => {
 
   describe('Price filter', () => {
     it('should display price filter button', async () => {
-      render(<SearchResultsContent />)
+      render(<SearchResults />)
       await act(async () => {})
 
       expect(screen.getByTestId('Prix')).toBeOnTheScreen()
     })
 
     it('should open the prices filter modal when pressing the prices filter button', async () => {
-      render(<SearchResultsContent />)
+      render(<SearchResults />)
       const priceButton = screen.getByTestId('Prix')
 
       await act(async () => {
@@ -249,7 +249,7 @@ describe('SearchResultsContent component', () => {
         ...mockSearchState,
         minPrice: '5',
       }
-      render(<SearchResultsContent />)
+      render(<SearchResults />)
       await act(async () => {})
 
       const priceButtonIcon = screen.getByTestId('priceButtonIcon')
@@ -275,14 +275,14 @@ describe('SearchResultsContent component', () => {
       })
 
       it('should display Duo filter button', async () => {
-        render(<SearchResultsContent />)
+        render(<SearchResults />)
         await act(async () => {})
 
         expect(screen.getByTestId('Duo')).toBeOnTheScreen()
       })
 
       it('should open the duo filter modal when pressing the duo filter button', async () => {
-        render(<SearchResultsContent />)
+        render(<SearchResults />)
         const duoButton = screen.getByTestId('Duo')
 
         await act(async () => {
@@ -311,7 +311,7 @@ describe('SearchResultsContent component', () => {
       })
 
       it('should not display Duo filter button', async () => {
-        render(<SearchResultsContent />)
+        render(<SearchResults />)
         await act(async () => {})
 
         expect(screen.queryByText('Duo')).not.toBeOnTheScreen()
@@ -330,7 +330,7 @@ describe('SearchResultsContent component', () => {
       })
 
       it('should not display Duo offer button', async () => {
-        render(<SearchResultsContent />)
+        render(<SearchResults />)
         await act(async () => {})
 
         expect(screen.queryByText('Duo')).not.toBeOnTheScreen()
@@ -349,7 +349,7 @@ describe('SearchResultsContent component', () => {
       })
 
       it('should not display Duo offer button', async () => {
-        render(<SearchResultsContent />)
+        render(<SearchResults />)
         await act(async () => {})
 
         expect(screen.queryByText('Duo')).not.toBeOnTheScreen()
@@ -359,7 +359,7 @@ describe('SearchResultsContent component', () => {
 
   describe('should not display geolocation incitation button', () => {
     it('when position is not null', async () => {
-      render(<SearchResultsContent />)
+      render(<SearchResults />)
       await act(async () => {})
 
       expect(screen.queryByText('Géolocalise-toi')).not.toBeOnTheScreen()
@@ -375,7 +375,7 @@ describe('SearchResultsContent component', () => {
         ...params,
       }
 
-      render(<SearchResultsContent />)
+      render(<SearchResults />)
       await act(async () => {})
 
       expect(screen.queryByText('Géolocalise-toi')).not.toBeOnTheScreen()
@@ -383,7 +383,7 @@ describe('SearchResultsContent component', () => {
 
     it('when position is null and no results search', async () => {
       mockPosition = null
-      render(<SearchResultsContent />)
+      render(<SearchResults />)
       await act(async () => {})
 
       expect(screen.queryByText('Géolocalise-toi')).not.toBeOnTheScreen()
@@ -393,7 +393,7 @@ describe('SearchResultsContent component', () => {
   describe('Location filter', () => {
     it('should display location filter button', async () => {
       mockPosition = null
-      render(<SearchResultsContent />)
+      render(<SearchResults />)
       await act(async () => {})
 
       expect(screen.getByTestId('Localisation')).toBeOnTheScreen()
@@ -419,7 +419,7 @@ describe('SearchResultsContent component', () => {
       }) => {
         mockPosition = position
         mockSearchState = { ...searchState, locationFilter }
-        render(<SearchResultsContent />)
+        render(<SearchResults />)
         await act(async () => {})
 
         expect(screen.queryByText(locationButtonLabel)).toBeOnTheScreen()
@@ -429,7 +429,7 @@ describe('SearchResultsContent component', () => {
 
   it(`should display ${venue.label} in location filter button label when a venue is selected`, async () => {
     mockSearchState = { ...searchState, venue }
-    render(<SearchResultsContent />)
+    render(<SearchResults />)
     await act(async () => {})
 
     expect(screen.queryByText(venue.label)).toBeOnTheScreen()
@@ -441,7 +441,7 @@ describe('SearchResultsContent component', () => {
     })
 
     it('should open the venue modal when pressing the venue filter button', async () => {
-      render(<SearchResultsContent />)
+      render(<SearchResults />)
 
       await act(async () => {
         const venueButton = screen.getByRole('button', { name: 'Lieu culturel' })
@@ -454,7 +454,7 @@ describe('SearchResultsContent component', () => {
     })
 
     it('should call set search state on press "Rechercher" in venue modal', async () => {
-      render(<SearchResultsContent />)
+      render(<SearchResults />)
 
       await act(async () => {
         const venueButton = screen.getByRole('button', { name: 'Lieu culturel' })
@@ -474,7 +474,7 @@ describe('SearchResultsContent component', () => {
     })
 
     it('when ENABLE_APP_LOCATION featureFlag, should display "Lieu culturel" in venue filter if no venue is selected', async () => {
-      render(<SearchResultsContent />)
+      render(<SearchResults />)
       await act(async () => {})
 
       expect(screen.getByTestId('venueButtonLabel')).toHaveTextContent('Lieu culturel')
@@ -485,7 +485,7 @@ describe('SearchResultsContent component', () => {
         ...searchState,
         venue,
       }
-      render(<SearchResultsContent />)
+      render(<SearchResults />)
       await act(async () => {})
 
       expect(screen.getByTestId('venueButtonLabel')).toHaveTextContent(venue.label)
@@ -496,7 +496,7 @@ describe('SearchResultsContent component', () => {
         ...searchState,
         locationFilter: { locationType: LocationMode.AROUND_ME, aroundRadius: MAX_RADIUS },
       }
-      render(<SearchResultsContent />)
+      render(<SearchResults />)
       await act(async () => {})
 
       expect(screen.getByTestId('venueButtonLabel')).toHaveTextContent('Lieu culturel')
@@ -507,7 +507,7 @@ describe('SearchResultsContent component', () => {
         ...searchState,
         locationFilter: { locationType: LocationMode.EVERYWHERE },
       }
-      render(<SearchResultsContent />)
+      render(<SearchResults />)
       await act(async () => {})
 
       expect(screen.getByTestId('venueButtonLabel')).toHaveTextContent('Lieu culturel')
@@ -516,14 +516,14 @@ describe('SearchResultsContent component', () => {
 
   describe('Dates and hours filter', () => {
     it('should display dates and hours filter button', async () => {
-      render(<SearchResultsContent />)
+      render(<SearchResults />)
       await act(async () => {})
 
       expect(screen.queryByTestId('Dates & heures')).toBeOnTheScreen()
     })
 
     it('should open the type filter modal when pressing the type filter button', async () => {
-      render(<SearchResultsContent />)
+      render(<SearchResults />)
       const datesHoursButton = screen.getByTestId('Dates & heures')
 
       await act(async () => {
@@ -546,7 +546,7 @@ describe('SearchResultsContent component', () => {
           ...mockSearchState,
           ...params,
         }
-        render(<SearchResultsContent />)
+        render(<SearchResults />)
         await act(async () => {})
 
         const datesHoursButtonIcon = screen.getByTestId('datesHoursButtonIcon')
@@ -567,7 +567,7 @@ describe('SearchResultsContent component', () => {
     mockPosition = null
     mockHits = mockedAlgoliaResponse.hits
     mockNbHits = mockedAlgoliaResponse.nbHits
-    render(<SearchResultsContent />)
+    render(<SearchResults />)
 
     await act(async () => {
       fireEvent.press(screen.getByText('Géolocalise-toi'))
@@ -578,38 +578,38 @@ describe('SearchResultsContent component', () => {
 
   it('should refetch results when user position received in a second time', async () => {
     mockPosition = null
-    render(<SearchResultsContent />)
+    render(<SearchResults />)
     await act(async () => {})
 
     expect(mockRefetch).not.toHaveBeenCalled()
 
     mockPosition = DEFAULT_POSITION
-    screen.rerender(<SearchResultsContent />)
+    screen.rerender(<SearchResults />)
 
     expect(mockRefetch).toHaveBeenCalledTimes(1)
   })
 
   it('should refetch results when user stop to share his position', async () => {
     mockPosition = DEFAULT_POSITION
-    render(<SearchResultsContent />)
+    render(<SearchResults />)
     await act(async () => {})
 
     // previousGeolocPosition is empty in first rendering
     expect(mockRefetch).toHaveBeenCalledTimes(1)
 
-    screen.rerender(<SearchResultsContent />)
+    screen.rerender(<SearchResults />)
 
     expect(mockRefetch).toHaveBeenCalledTimes(1)
 
     mockPosition = null
-    screen.rerender(<SearchResultsContent />)
+    screen.rerender(<SearchResults />)
 
     // first rendering + rendering when user stop to share his position
     expect(mockRefetch).toHaveBeenCalledTimes(2)
   })
 
   it('should not log PerformSearch when there is not search query execution', async () => {
-    render(<SearchResultsContent />)
+    render(<SearchResults />)
     await act(async () => {})
 
     expect(analytics.logPerformSearch).not.toHaveBeenCalled()
@@ -617,37 +617,37 @@ describe('SearchResultsContent component', () => {
 
   it('should log PerformSearch only one time when there is search query execution and several re-render', async () => {
     mockIsLoading = true
-    render(<SearchResultsContent />)
+    render(<SearchResults />)
     await act(async () => {})
 
     expect(analytics.logPerformSearch).not.toHaveBeenCalled()
 
     mockIsLoading = false
-    screen.rerender(<SearchResultsContent />)
+    screen.rerender(<SearchResults />)
 
     expect(analytics.logPerformSearch).toHaveBeenCalledTimes(1)
 
-    screen.rerender(<SearchResultsContent />)
+    screen.rerender(<SearchResults />)
 
     expect(analytics.logPerformSearch).toHaveBeenCalledTimes(1)
   })
 
   it('should log PerformSearch with search result when there is search query execution', async () => {
     mockIsLoading = true
-    render(<SearchResultsContent />)
+    render(<SearchResults />)
     await act(async () => {})
 
     expect(analytics.logPerformSearch).not.toHaveBeenCalled()
 
     mockIsLoading = false
     mockSearchState = searchState
-    screen.rerender(<SearchResultsContent />)
+    screen.rerender(<SearchResults />)
 
     expect(analytics.logPerformSearch).toHaveBeenNthCalledWith(1, mockSearchState, mockNbHits)
   })
 
   it('should not log NoSearchResult when there is not search query execution', async () => {
-    render(<SearchResultsContent />)
+    render(<SearchResults />)
     await act(async () => {})
 
     expect(analytics.logNoSearchResult).not.toHaveBeenCalled()
@@ -655,31 +655,31 @@ describe('SearchResultsContent component', () => {
 
   it('should log NoSearchResult only one time when there is search query execution, nbHits = 0 and several re-render', async () => {
     mockIsLoading = true
-    render(<SearchResultsContent />)
+    render(<SearchResults />)
     await act(async () => {})
 
     expect(analytics.logNoSearchResult).not.toHaveBeenCalled()
 
     mockIsLoading = false
-    screen.rerender(<SearchResultsContent />)
+    screen.rerender(<SearchResults />)
 
     expect(analytics.logNoSearchResult).toHaveBeenCalledTimes(1)
 
-    screen.rerender(<SearchResultsContent />)
+    screen.rerender(<SearchResults />)
 
     expect(analytics.logNoSearchResult).toHaveBeenCalledTimes(1)
   })
 
   it('should log NoSearchResult with search result when there is search query execution and nbHits = 0', async () => {
     mockIsLoading = true
-    render(<SearchResultsContent />)
+    render(<SearchResults />)
     await act(async () => {})
 
     expect(analytics.logNoSearchResult).not.toHaveBeenCalled()
 
     mockIsLoading = false
     mockSearchState = searchState
-    screen.rerender(<SearchResultsContent />)
+    screen.rerender(<SearchResults />)
 
     expect(analytics.logNoSearchResult).toHaveBeenNthCalledWith(1, '', searchId)
   })
@@ -688,7 +688,7 @@ describe('SearchResultsContent component', () => {
     mockPosition = null
     mockHits = mockedAlgoliaResponse.hits
     mockNbHits = mockedAlgoliaResponse.nbHits
-    render(<SearchResultsContent />)
+    render(<SearchResults />)
 
     await act(async () => {
       fireEvent.press(screen.getByText('Géolocalise-toi'))
@@ -705,7 +705,7 @@ describe('SearchResultsContent component', () => {
     })
 
     it('when position is null', async () => {
-      render(<SearchResultsContent />)
+      render(<SearchResults />)
       await act(async () => {})
 
       expect(screen.queryByText('Géolocalise-toi')).toBeOnTheScreen()
@@ -714,7 +714,7 @@ describe('SearchResultsContent component', () => {
     it('when position is null and query is not an offer not present', async () => {
       mockUserData = [{ message: 'n’est pas disponible sur le pass Culture.' }]
       mockSearchState = { ...searchState, query: 'iPhone' }
-      render(<SearchResultsContent />)
+      render(<SearchResults />)
       await act(async () => {})
 
       expect(screen.queryByText('Géolocalise-toi')).not.toBeOnTheScreen()
@@ -727,7 +727,7 @@ describe('SearchResultsContent component', () => {
       mockNbHits = mockedAlgoliaResponse.nbHits
       mockUserData = [{ message: 'Offre non disponible sur le pass Culture.' }]
       mockSearchState = { ...searchState, query: 'iPhone' }
-      render(<SearchResultsContent />)
+      render(<SearchResults />)
       await act(async () => {})
 
       expect(screen.queryByText('Offre non disponible sur le pass Culture.')).toBeOnTheScreen()
@@ -738,7 +738,7 @@ describe('SearchResultsContent component', () => {
       mockNbHits = mockedAlgoliaResponse.nbHits
       mockUserData = []
       mockSearchState = { ...searchState, query: 'Deezer' }
-      render(<SearchResultsContent />)
+      render(<SearchResults />)
       await act(async () => {})
 
       expect(screen.queryByText('Offre non disponible sur le pass Culture.')).not.toBeOnTheScreen()
@@ -753,7 +753,7 @@ describe('SearchResultsContent component', () => {
         offerIsDuo: true,
       }
 
-      render(<SearchResultsContent />)
+      render(<SearchResults />)
       let filterButton
       await act(async () => {
         filterButton = screen.getByLabelText('Voir tous les filtres\u00a0: 2 filtres actifs')
