--- src/features/search/pages/SearchLanding/SearchLanding.native.test.tsx
@@ -1,474 +1,0 @@
-import mockdate from 'mockdate'
-import React from 'react'
-import { Keyboard } from 'react-native'
-import { v4 as uuidv4 } from 'uuid'
-
-import { useRoute } from '__mocks__/@react-navigation/native'
-import { NativeCategoryIdEnumv2, SearchGroupNameEnumv2 } from 'api/gen'
-import { initialSearchState } from 'features/search/context/reducer'
-import { mockedSearchHistory } from 'features/search/fixtures/mockedSearchHistory'
-import * as useFilterCountAPI from 'features/search/helpers/useFilterCount/useFilterCount'
-import { SearchLanding } from 'features/search/pages/SearchLanding/SearchLanding'
-import { SearchState, SearchView } from 'features/search/types'
-import { Venue } from 'features/venue/types'
-import { analytics } from 'libs/analytics'
-import { env } from 'libs/environment'
-import { LocationMode } from 'libs/location/types'
-import { useNetInfoContext as useNetInfoContextDefault } from 'libs/network/NetInfoWrapper'
-import { SuggestedPlace } from 'libs/place'
-import { placeholderData } from 'libs/subcategories/placeholderData'
-import { mockedSuggestedVenues } from 'libs/venue/fixtures/mockedSuggestedVenues'
-import { act, fireEvent, render, screen } from 'tests/utils'
-
-const venue: Venue = mockedSuggestedVenues[0]
-
-let mockSearchState: SearchState = {
-  ...initialSearchState,
-  offerCategories: [SearchGroupNameEnumv2.FILMS_SERIES_CINEMA],
-  venue,
-  priceRange: [0, 20],
-}
-
-const mockDispatch = jest.fn()
-let mockIsFocusOnSuggestions = false
-jest.mock('features/search/context/SearchWrapper', () => ({
-  useSearch: () => ({
-    searchState: mockSearchState,
-    dispatch: mockDispatch,
-    isFocusOnSuggestions: mockIsFocusOnSuggestions,
-    hideSuggestions: jest.fn(),
-  }),
-}))
-
-jest.mock('react-query')
-
-const mockData = { pages: [{ nbHits: 0, hits: [], page: 0 }] }
-const mockHasNextPage = true
-const mockFetchNextPage = jest.fn()
-
-jest.mock('features/search/api/useSearchResults/useSearchResults', () => ({
-  useSearchResults: () => ({
-    data: mockData,
-    hits: [],
-    nbHits: 0,
-    isFetching: false,
-    isLoading: false,
-    hasNextPage: mockHasNextPage,
-    fetchNextPage: mockFetchNextPage,
-    isFetchingNextPage: false,
-  }),
-}))
-
-jest.mock('libs/network/useNetInfo', () => jest.requireMock('@react-native-community/netinfo'))
-const mockUseNetInfoContext = useNetInfoContextDefault as jest.Mock
-
-const mockSettings = jest.fn().mockReturnValue({ data: {} })
-jest.mock('features/auth/context/SettingsContext', () => ({
-  useSettingsContext: jest.fn(() => mockSettings()),
-}))
-jest.mock('features/search/helpers/useSearchHistory/useSearchHistory', () => ({
-  useSearchHistory: () => ({
-    filteredHistory: [],
-    addToHistory: jest.fn(),
-    removeFromHistory: jest.fn(),
-    search: jest.fn(),
-  }),
-}))
-
-const mockHits = [
-  {
-    objectID: '1',
-    offer: { name: 'Test1', searchGroupName: 'MUSIQUE' },
-    _highlightResult: {
-      query: {
-        value: '<mark>Test1</mark>',
-        matchLevel: 'full',
-        fullyHighlighted: true,
-        matchedWords: ['Test1'],
-      },
-    },
-    [env.ALGOLIA_OFFERS_INDEX_NAME]: {
-      exact_nb_hits: 2,
-      facets: {
-        analytics: {
-          ['offer.searchGroupNamev2']: [
-            {
-              attribute: '',
-              operator: '',
-              value: SearchGroupNameEnumv2.FILMS_SERIES_CINEMA,
-              count: 10,
-            },
-          ],
-          ['offer.nativeCategoryId']: [
-            {
-              attribute: '',
-              operator: '',
-              value: NativeCategoryIdEnumv2.SEANCES_DE_CINEMA,
-              count: 10,
-            },
-          ],
-        },
-      },
-    },
-  },
-  {
-    objectID: '2',
-    offer: { name: 'Test2', searchGroupName: 'MUSIQUE' },
-    _geoloc: {},
-    _highlightResult: {
-      query: {
-        value: '<mark>Test2</mark>',
-        matchLevel: 'full',
-        fullyHighlighted: true,
-        matchedWords: ['Test2'],
-      },
-    },
-    [env.ALGOLIA_OFFERS_INDEX_NAME]: {
-      exact_nb_hits: 2,
-      facets: {
-        analytics: {
-          ['offer.searchGroupNamev2']: [
-            {
-              attribute: '',
-              operator: '',
-              value: SearchGroupNameEnumv2.FILMS_SERIES_CINEMA,
-              count: 10,
-            },
-          ],
-          ['offer.nativeCategoryId']: [
-            {
-              attribute: '',
-              operator: '',
-              value: NativeCategoryIdEnumv2.SEANCES_DE_CINEMA,
-              count: 10,
-            },
-          ],
-        },
-      },
-    },
-  },
-]
-
-jest.mock('react-instantsearch-core', () => ({
-  ...jest.requireActual('react-instantsearch-core'),
-  useSearchBox: () => ({
-    query: '',
-    refine: jest.fn,
-  }),
-  useInfiniteHits: () => ({
-    hits: mockHits,
-  }),
-}))
-
-jest.spyOn(useFilterCountAPI, 'useFilterCount').mockReturnValue(3)
-jest.mock('algoliasearch')
-
-const mockSubcategoriesData = placeholderData
-jest.mock('libs/subcategories/useSubcategories', () => ({
-  useSubcategories: () => ({
-    data: mockSubcategoriesData,
-  }),
-}))
-
-const TODAY_DATE = new Date('2023-09-25T00:00:00.000Z')
-
-const mockUseSearchHistory = jest.fn()
-jest.mock('features/search/helpers/useSearchHistory/useSearchHistory', () => ({
-  useSearchHistory: jest.fn(() => mockUseSearchHistory()),
-}))
-mockUseSearchHistory.mockReturnValue({
-  filteredHistory: mockedSearchHistory,
-  queryHistory: '',
-  addToHistory: jest.fn(),
-  removeFromHistory: jest.fn(),
-  search: jest.fn(),
-})
-
-const searchId = uuidv4()
-
-const mockedPlace: SuggestedPlace = {
-  label: 'Kourou',
-  info: 'Guyane',
-  geolocation: { longitude: -52.669736, latitude: 5.16186 },
-}
-
-const mockSetPlace = jest.fn()
-const mockSetSelectedLocationMode = jest.fn()
-let mockHasGeolocPosition = false
-
-jest.mock('libs/location/LocationWrapper', () => ({
-  useLocation: () => ({
-    setPlace: mockSetPlace,
-    place: mockedPlace,
-    onModalHideRef: jest.fn(),
-    isCurrentLocationMode: jest.fn(),
-    setSelectedLocationMode: mockSetSelectedLocationMode,
-    hasGeolocPosition: mockHasGeolocPosition,
-  }),
-}))
-
-describe('<SearchLanding />', () => {
-  mockUseNetInfoContext.mockReturnValue({ isConnected: true })
-
-  afterEach(() => {
-    mockSearchState = {
-      ...initialSearchState,
-      offerCategories: [SearchGroupNameEnumv2.FILMS_SERIES_CINEMA],
-      venue,
-      priceRange: [0, 20],
-    }
-    mockIsFocusOnSuggestions = false
-  })
-
-  it('should render SearchLanding', async () => {
-    render(<SearchLanding />)
-
-    await screen.findByText('Rechercher')
-
-    await act(() => {})
-
-    expect(screen).toMatchSnapshot()
-  })
-
-  it('should setPlace and setLocationMode in location context, when URI params contains a place,', async () => {
-    useRoute.mockReturnValueOnce({
-      params: {
-        locationFilter: {
-          locationType: LocationMode.AROUND_PLACE,
-          place: mockedPlace,
-        },
-      },
-    })
-
-    render(<SearchLanding />)
-
-    await act(async () => {})
-
-    expect(mockSetPlace).toHaveBeenCalledWith(mockedPlace)
-    expect(mockSetSelectedLocationMode).toHaveBeenCalledWith(LocationMode.AROUND_PLACE)
-  })
-
-  it('should setLocationMode to AROUND-ME in location context,when URI params contains AROUND-ME and hasGeolocPosition is true', async () => {
-    mockHasGeolocPosition = true
-    useRoute.mockReturnValueOnce({
-      params: {
-        locationFilter: {
-          locationType: LocationMode.AROUND_ME,
-        },
-      },
-    })
-
-    render(<SearchLanding />)
-
-    await act(async () => {})
-
-    expect(mockSetSelectedLocationMode).toHaveBeenCalledWith(LocationMode.AROUND_ME)
-  })
-
-  it("shouldn't setLocationMode to AROUND-ME in location context,when URI params contains AROUND-ME and hasGeolocPosition is false", async () => {
-    mockHasGeolocPosition = false
-
-    useRoute.mockReturnValueOnce({
-      params: {
-        locationFilter: {
-          locationType: LocationMode.AROUND_ME,
-        },
-      },
-    })
-
-    render(<SearchLanding />)
-
-    await act(async () => {})
-
-    expect(mockSetSelectedLocationMode).not.toHaveBeenCalledWith(LocationMode.AROUND_ME)
-  })
-
-  describe('When SearchLanding is focus on suggestions', () => {
-    beforeEach(() => {
-      mockIsFocusOnSuggestions = true
-    })
-
-    it('should display offer suggestions', async () => {
-      render(<SearchLanding />)
-      await act(async () => {})
-
-      expect(screen.getByTestId('autocompleteOfferItem_1')).toBeOnTheScreen()
-      expect(screen.getByTestId('autocompleteOfferItem_2')).toBeOnTheScreen()
-    })
-
-    it('should display venue suggestions', async () => {
-      render(<SearchLanding />)
-      await act(async () => {})
-
-      expect(screen.getByTestId('autocompleteVenueItem_1')).toBeOnTheScreen()
-      expect(screen.getByTestId('autocompleteVenueItem_2')).toBeOnTheScreen()
-    })
-
-    it('should handle venue press', async () => {
-      render(<SearchLanding />)
-      await act(async () => {})
-
-      expect(screen.getByTestId('autocompleteVenueItem_1')).toBeOnTheScreen()
-
-      fireEvent.press(screen.getByTestId('autocompleteVenueItem_1'))
-
-      expect(analytics.logConsultVenue).toHaveBeenCalledWith({
-        from: 'searchAutoComplete',
-        venueId: 1,
-      })
-    })
-
-    it('should dismiss keyboard on scroll', async () => {
-      const scrollEventBottom = {
-        nativeEvent: {
-          layoutMeasurement: { height: 1000 },
-          contentOffset: { y: 900 },
-          contentSize: { height: 1600 },
-        },
-      }
-      const keyboardDismissSpy = jest.spyOn(Keyboard, 'dismiss')
-      render(<SearchLanding />)
-      await act(async () => {})
-
-      const scrollView = screen.getByTestId('autocompleteScrollView')
-      // 1st scroll to bottom => trigger
-      scrollView.props.onScroll(scrollEventBottom)
-
-      expect(keyboardDismissSpy).toHaveBeenCalledTimes(1)
-    })
-
-    it('should display search history when it has items', async () => {
-      mockdate.set(TODAY_DATE)
-      render(<SearchLanding />)
-      await act(async () => {})
-
-      expect(screen.getByText('Historique de recherche')).toBeOnTheScreen()
-    })
-
-    it('should not display search history when it has not items', async () => {
-      mockdate.set(TODAY_DATE)
-      mockUseSearchHistory
-        .mockReturnValueOnce({
-          filteredHistory: [],
-          queryHistory: '',
-          addToHistory: jest.fn(),
-          removeFromHistory: jest.fn(),
-          search: jest.fn(),
-        })
-        .mockReturnValueOnce({
-          filteredHistory: [],
-          queryHistory: '',
-          addToHistory: jest.fn(),
-          removeFromHistory: jest.fn(),
-          search: jest.fn(),
-        })
-        .mockReturnValueOnce({
-          filteredHistory: [],
-          queryHistory: '',
-          addToHistory: jest.fn(),
-          removeFromHistory: jest.fn(),
-          search: jest.fn(),
-        })
-      render(<SearchLanding />)
-      await act(async () => {})
-
-      expect(screen.queryByText('Historique de recherche')).not.toBeOnTheScreen()
-    })
-
-    it('should dismiss the keyboard when pressing search history item', async () => {
-      mockdate.set(TODAY_DATE)
-      const keyboardDismissSpy = jest.spyOn(Keyboard, 'dismiss')
-      render(<SearchLanding />)
-      await act(async () => {})
-
-      fireEvent.press(screen.getByText('manga'))
-
-      expect(keyboardDismissSpy).toHaveBeenCalledTimes(1)
-    })
-
-    describe('should update state and execute the search with the history item', () => {
-      it('When it has not category and native category', async () => {
-        mockdate.set(TODAY_DATE)
-        render(<SearchLanding />)
-        await act(async () => {})
-
-        fireEvent.press(screen.getByText('manga'))
-
-        expect(mockDispatch).toHaveBeenCalledWith({
-          type: 'SET_STATE',
-          payload: {
-            ...mockSearchState,
-            query: 'manga',
-            view: SearchView.Results,
-            searchId,
-            isFromHistory: true,
-            isAutocomplete: undefined,
-            offerGenreTypes: undefined,
-            offerNativeCategories: undefined,
-            offerCategories: [],
-          },
-        })
-      })
-
-      it('When it has category and native category', async () => {
-        mockdate.set(TODAY_DATE)
-        render(<SearchLanding />)
-        await act(async () => {})
-
-        fireEvent.press(screen.getByText('tolkien'))
-
-        expect(mockDispatch).toHaveBeenCalledWith({
-          type: 'SET_STATE',
-          payload: {
-            ...mockSearchState,
-            query: 'tolkien',
-            view: SearchView.Results,
-            searchId,
-            isFromHistory: true,
-            isAutocomplete: undefined,
-            offerGenreTypes: undefined,
-            offerNativeCategories: [NativeCategoryIdEnumv2.LIVRES_AUDIO_PHYSIQUES],
-            offerCategories: [SearchGroupNameEnumv2.LIVRES],
-          },
-        })
-      })
-
-      it('When it has only a category', async () => {
-        mockdate.set(TODAY_DATE)
-        render(<SearchLanding />)
-        await act(async () => {})
-
-        fireEvent.press(screen.getByText('foresti'))
-
-        expect(mockDispatch).toHaveBeenCalledWith({
-          type: 'SET_STATE',
-          payload: {
-            ...mockSearchState,
-            query: 'foresti',
-            view: SearchView.Results,
-            searchId,
-            isFromHistory: true,
-            isAutocomplete: undefined,
-            offerGenreTypes: undefined,
-            offerNativeCategories: undefined,
-            offerCategories: [SearchGroupNameEnumv2.SPECTACLES],
-          },
-        })
-      })
-    })
-  })
-
-  describe('When offline', () => {
-    it('should display offline page', async () => {
-      mockUseNetInfoContext
-        .mockReturnValueOnce({ isConnected: false })
-        .mockReturnValueOnce({ isConnected: false })
-        .mockReturnValueOnce({ isConnected: false })
-      render(<SearchLanding />)
-      await act(async () => {})
-      await act(async () => {})
-
-      expect(screen.getByText('Pas de réseau internet')).toBeOnTheScreen()
-    })
-  })
-})
