--- src/features/search/pages/SearchLanding/SearchLanding.web.test.tsx
@@ -1,114 +1,0 @@
-import React from 'react'
-
-import { SearchGroupNameEnumv2 } from 'api/gen'
-import { initialSearchState } from 'features/search/context/reducer'
-import * as useFilterCountAPI from 'features/search/helpers/useFilterCount/useFilterCount'
-import { SearchLanding } from 'features/search/pages/SearchLanding/SearchLanding'
-import { SearchState } from 'features/search/types'
-import { Venue } from 'features/venue/types'
-import * as useFeatureFlagAPI from 'libs/firebase/firestore/featureFlags/useFeatureFlag'
-import { useNetInfoContext as useNetInfoContextDefault } from 'libs/network/NetInfoWrapper'
-import { mockedSuggestedVenues } from 'libs/venue/fixtures/mockedSuggestedVenues'
-import { act, checkAccessibilityFor, render } from 'tests/utils/web'
-
-const venue: Venue = mockedSuggestedVenues[0]
-
-const mockSearchState: SearchState = {
-  ...initialSearchState,
-  offerCategories: [SearchGroupNameEnumv2.FILMS_SERIES_CINEMA],
-  venue,
-  priceRange: [0, 20],
-}
-
-jest.mock('features/search/context/SearchWrapper', () => ({
-  useSearch: () => ({ searchState: mockSearchState, dispatch: jest.fn() }),
-}))
-
-jest.mock('features/auth/context/AuthContext')
-
-jest.mock('react-query')
-
-jest.mock('features/search/api/useSearchResults/useSearchResults', () => ({
-  useSearchResults: () => ({
-    data: { pages: [{ nbHits: 0, hits: [], page: 0 }] },
-    hits: [],
-    nbHits: 0,
-    isFetching: false,
-    isLoading: false,
-    hasNextPage: true,
-    fetchNextPage: jest.fn(),
-    isFetchingNextPage: false,
-  }),
-}))
-
-jest.mock('libs/network/useNetInfo', () => jest.requireMock('@react-native-community/netinfo'))
-const mockUseNetInfoContext = useNetInfoContextDefault as jest.Mock
-
-const mockSettings = jest.fn().mockReturnValue({ data: {} })
-jest.mock('features/auth/context/SettingsContext', () => ({
-  useSettingsContext: jest.fn(() => mockSettings()),
-}))
-
-jest.mock('react-instantsearch-core', () => ({
-  ...jest.requireActual('react-instantsearch-core'),
-  useSearchBox: () => ({
-    query: '',
-    refine: jest.fn,
-  }),
-  useInfiniteHits: () => ({
-    hits: [
-      {
-        objectID: '1',
-        offer: { name: 'Test1', searchGroupName: 'MUSIQUE' },
-        _geoloc: {},
-      },
-      {
-        objectID: '2',
-        offer: { name: 'Test2', searchGroupName: 'MUSIQUE' },
-        _geoloc: {},
-      },
-    ],
-  }),
-}))
-
-jest.spyOn(useFilterCountAPI, 'useFilterCount').mockReturnValue(3)
-jest.mock('algoliasearch')
-
-const mockV4 = jest.fn()
-jest.mock('uuid', () => ({
-  v1: jest.fn(),
-  v4: jest.fn(mockV4),
-}))
-
-// mockUseNetInfoContext.mockReturnValue({ isConnected: true })
-jest.spyOn(useFeatureFlagAPI, 'useFeatureFlag').mockReturnValue(true)
-
-describe('<SearchLanding />', () => {
-  describe('Accessibility', () => {
-    it('should not have basic accessibility issues', async () => {
-      mockUseNetInfoContext.mockReturnValueOnce({ isConnected: true })
-      const { container } = render(<SearchLanding />)
-
-      await act(async () => {})
-
-      await act(async () => {
-        const results = await checkAccessibilityFor(container)
-
-        expect(results).toHaveNoViolations()
-      })
-    })
-
-    it('should not have basic accessibility issues when offline', async () => {
-      mockUseNetInfoContext.mockReturnValueOnce({ isConnected: false })
-      const { container } = render(<SearchLanding />)
-
-      await act(async () => {})
-
-      await act(async () => {
-        const results = await checkAccessibilityFor(container)
-
-        expect(results).toHaveNoViolations()
-      })
-    })
-  })
-})
